% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msod_advi.R
\name{coef.msod_advi}
\alias{coef.msod_advi}
\title{Retrieve draws for coefficients in MSOD model}
\usage{
\method{coef}{msod_advi}(obj)
}
\arguments{
\item{obj}{The fitted MSOD model.}
}
\value{
A list containing the following fields: (1) `env_intercepts`, a
DataFrame of draws of the suitability intercepts for each species; (2)
`env_slopes`, a list mapping from species names to DataFrames of draws for
each suitability covariate; (3) `obs_slopes`, a list mapping from species
names to DataFrames of draws for each observation covariate; (4)
obs_prior_means, draws from the group prior means estimated for the
observation process; (5) obs_prior_sds, draws from the group prior standard
deviations for the observation process.
}
\description{
Returns draws of the estimated parameters.
}
