% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssod_ml.R
\name{coef.ssod_maxlik}
\alias{coef.ssod_maxlik}
\title{Extract coefficients of fitted SSOD model}
\usage{
\method{coef}{ssod_maxlik}(obj)
}
\arguments{
\item{obj}{The fitted SSOD model.}
}
\value{
A list with two names: `env_coefs`, containing the coefficients
driving suitability (often referred to by the Greek letter Psi), and
`obs_coefs`, containing the coefficients driving observation.
}
\description{
Extract coefficients of fitted SSOD model
}
