% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_data.R
\docType{data}
\name{eBird}
\alias{eBird}
\title{Subset of eBird data}
\format{
A list.
}
\source{
\href{https://ebird.org/science/use-ebird-data/download-ebird-data-products}{eBird
    Basic Dataset}
}
\usage{
data(eBird)
}
\description{
Data from eBird (\href{https://ebird.org/home}{eBird}) for 32 species,
collected in June 2019. The data is a random sample of 8,000 checklists and
is meant to illustrate the format used by the multi-species occupancy
detection models in this package.
}
\references{
eBird. 2021. eBird: An online database of bird distribution and
    abundance [web application]. eBird, Cornell Lab of Ornithology, Ithaca,
    New York. Available: \href{http://www.ebird.org}{eBird}.
}
\keyword{datasets}
