% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{install_occu_py}
\alias{install_occu_py}
\title{Installs the python requirements of the package.}
\usage{
install_occu_py(gpu = TRUE)
}
\arguments{
\item{gpu}{Whether or not to install the GPU version of the package. Note
that using the GPU version is highly recommended.}
}
\description{
Installs the python requirements of the package.
}
