% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msod_advi.R
\name{msod_vi}
\alias{msod_vi}
\title{Fit MSOD model}
\usage{
msod_vi(
  env_formula,
  obs_formula,
  X_env,
  X_checklist,
  y_checklist,
  checklist_cell_ids,
  M = 20L,
  n_draws = 1000L,
  verbose = TRUE
)
}
\arguments{
\item{env_formula}{Formula to use for the environment / suitability part of
the model. Should be a string compatible with the patsy python package.}

\item{obs_formula}{Formula to use for the observation / detection part of the
model. Again, should be supported by patsy.}

\item{X_env}{A DataFrame containing the environmental covariates. Dimensions
(n_sites, n_env_covariates).}

\item{X_checklist}{A DataFrame containing the observation
covariates. Dimensions (n_checklists, n_obs_covariates).}

\item{y_checklist}{A DataFrame specifying whether each species was or was not
observed for each visit (checklist). Dimensions (n_checklists, n_species).}

\item{checklist_cell_ids}{A vector specifying which site (aka cell) each
checklist belongs to. E.g. if checklist_cell_ids[2] = 3, the second checklist
was made at site 3. Note that sites are number from 0, so this would be the
_fourth_ site!}

\item{M}{The number of fixed draws to use in the variational
approximations. Should be set as large as possible, though M > 100 is
unlikely to change results.}

\item{n_draws}{The number of draws to make from the approximate posterior.}

\item{verbose}{Whether or not to print the progress of the optimisation.}
}
\value{
The fitted msod_advi object.
}
\description{
Fits a multi-species occupancy detection (MSOD) model using variational
inference.
}
