% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msod_advi.R
\name{predict.msod_advi}
\alias{predict.msod_advi}
\title{Predict using MSOD model}
\usage{
\method{predict}{msod_advi}(obj, X_env_new, X_obs_new = NULL, type = "env")
}
\arguments{
\item{obj}{The fitted MSOD model.}

\item{X_env_new}{The new environmental covariates to predict suitability for.}

\item{X_obs_new}{The new observation covariates to use for prediction. Can be
NULL if only environmental suitability is of interest.}

\item{type}{The type of prediction to make. If type = 'env', the probability
of presence is returned, and X_obs_new can be NULL. If type = 'obs', the
probability of detection is returned (i.e., species is present _and_
detected), and X_obs_new must be provided.}
}
\value{
A DataFrame with one column for each species, filled with the
    predicted probabilities.
}
\description{
Predicts the probability that a site is suitable, or that a species is
detected, given environmental and detection covariates.
}
