% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msod_advi.R
\name{restore_model}
\alias{restore_model}
\title{Restore the fitted MSOD model from disk}
\usage{
restore_model(save_dir)
}
\arguments{
\item{save_dir}{The directory the MSOD model was saved to.}
}
\description{
Restore the fitted MSOD model from disk
}
