% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msod_advi.R
\name{save_model}
\alias{save_model}
\title{Save the fitted MSOD model to disk}
\usage{
save_model(obj, save_dir)
}
\arguments{
\item{obj}{The fitted MSOD model object.}

\item{save_dir}{The folder to save the model to.}
}
\description{
Save the fitted MSOD model to disk
}
