% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{set_gpu}
\alias{set_gpu}
\title{Set whether to use the GPU or not}
\usage{
set_gpu(use_gpu = TRUE)
}
\arguments{
\item{use_gpu}{Whether or not to use the GPU.}
}
\description{
If use of a GPU is desired (and this is highly recommended), this function
must be called at the start of the program.
}
