% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssod_ml.R
\name{summary.ssod_maxlik}
\alias{summary.ssod_maxlik}
\title{Summarise the SSOD model}
\usage{
\method{summary}{ssod_maxlik}(obj)
}
\arguments{
\item{obj}{The fitted SSOD model}
}
\value{
Nothing, but prints the coefficients and the gradient norm at the
optimum. If the gradient norm is large, this may signal lack of
convergence. Ideally, it should be as close to zero as possible.
}
\description{
Summarise the SSOD model
}
